Criando arquivo object para userInputReader
