+++ solutions/devsprint-luiz-zytkowski-1/src/main/kotlin/framework/exceptions/EmptyUserNameException.kt
@@ -1,0 +1,5 @@
+package framework.exceptions
+
+import framework.OwnedException
+
+class EmptyUserNameException(message: String) : OwnedException(message)
