+++ solutions/devsprint-luiz-zytkowski-1/src/main/kotlin/framework/exceptions/InvalidUserNameException.kt
@@ -1,0 +1,5 @@
+package framework.exceptions
+
+import framework.OwnedException
+
+class InvalidUserNameException(message: String) : OwnedException(message)
